requirements.txt
asgiref==3.6.0
cffi==1.15.1
colorama==0.4.6
cryptography==39.0.0
Django==4.1.5
django-crispy-forms==1.14.0
django-qr-code==3.1.1
Pillow==9.4.0
gunicorn==20.1.0
pycparser==2.21
pydantic==1.10.4
python-barcode==0.14.0
python-dotenv==0.21.0
pytz==2022.7.1
qrcode==7.3.1
segno==1.5.2
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.7
